requirements.txt
streamlit
pandas
networkx
pydantic
python-dotenv
requests
ollama
sentence-transformers
faiss-cpu
pypdf